Delete unused files and add About component
