+++ app/(tabs)/about.tsx
@@ -1,49 +1,29 @@
-import { ScrollView, StyleSheet } from 'react-native';
+import { View, StyleSheet, ImageBackground } from 'react-native';
+import About from '../../components/forest/About';
+import React from 'react';
 
-import AboutCardWithImage from '../../components/leaf/AboutCardWithImage';
-import { View } from 'react-native';
-export default function About() {
+export default function AboutPage() {
   return (
-    <ScrollView contentContainerStyle={{ flexGrow: 1 }} keyboardShouldPersistTaps='handled' scrollEnabled={true}>
-      <View style={styles.container}>
-        <View style={styles.aboutCardTop}>
-          <AboutCardWithImage
-            imageName='icon_negate'
-            title='Time is Money'
-            text='テストテストテストテストテストテストテストテストテストテストテストテストテストテストテストテストテストテストテストテストテストテストテストテストテストテストテストテストテストテストテストテストテストテストテストテストテストテストテストテストテストテストテストテストテストテストテストテストテストテストテストテストテストテスト'
-          />
-        </View>
-        <View style={styles.aboutCardNonTop}>
-          <AboutCardWithImage
-            imageName='icon_negate'
-            title='Time is Money'
-            text='テストテストテストテストテストテストテストテストテストテストテストテストテストテストテストテストテストテスト'
-          />
-        </View>
-        <View style={styles.aboutCardNonTop}>
-          <AboutCardWithImage
-            imageName='icon_negate'
-            title='Time is Money'
-            text='テストテストテストテストテストテストテストテストテストテストテストテストテストテストテストテストテストテスト'
-          />
-        </View>
-      </View>
-    </ScrollView>
+    <View style={styles.container}>
+      <ImageBackground
+        source={require('../../assets/images/about_background_image.png')}
+        style={styles.backgroundImage}
+      >
+        <About />
+      </ImageBackground>
+    </View>
   );
 }
 
 const styles = StyleSheet.create({
   container: {
-    alignItems: 'center',
-    justifyContent: 'center',
-    padding: 25,
-    backgroundColor: 'white',
+    flex: 1,
   },
-  aboutCardTop: {
+  backgroundImage: {
+    position: 'absolute',
+    top: 0,
+    left: 0,
     width: '100%',
-  },
-  aboutCardNonTop: {
-    width: '100%',
-    marginTop: 30,
+    height: '100%',
   },
 });
